Fixing warning: "This expression will be resolved to Int in future releases. Please add explicit conversion call"

+++ phoenix-shared/src/commonMain/kotlin/fr.acinq.phoenix/controllers/payments/ReceiveController.kt
@@ -56,7 +56,7 @@ class AppReceiveController(
                                 paymentPreimage = preimage,
                                 amount = intent.amount,
                                 description = intent.description,
-                                expirySeconds = 3600 * 24 * 7, // one week
+                                expirySeconds = (3600 * 24 * 7).toLong(), // one week
                                 result = deferred
                             )
                         )
